fix: instantiate wagmi client on boot, autologin
